Refactor MpeContainer and MpeFormContainer
